backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "veloxship-backend"
version = "0.1.0"
description = "Logistics Delivery Bill Management — Backend API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "sqlalchemy>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "python-multipart>=0.0.6",
    "weasyprint>=61",
    "qrcode>=7.4",
    "python-barcode>=0.15",
    "unidecode>=1.3",
    "structlog>=24.0",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "testcontainers[postgres]>=4.0",
    "ruff>=0.3",
    "black>=24.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
